# docs/swagger.yml
---
openapi: 3.1.0
info:
  title: API Documentation
  description: kotlin springboot sample project입니다.
  version: 1.0.0
servers:
- url: "/"
  description: Default server
tags:
- name: Users
  description: 회원가입 API
- name: Orders
  description: 주문 조회와 상태 전이 API
paths:
  "/signup":
    post:
      tags:
      - Users
      summary: 회원가입
      description: 새 사용자를 등록합니다.
      operationId: signUp
      requestBody:
        description: 회원가입 요청 바디
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SignupRequest"
            examples:
              request:
                "$ref": "#/components/examples/SignupRequest"
        required: true
      responses:
        '201':
          description: 회원가입 성공
          content:
            application/json:
              examples:
                success:
                  "$ref": "#/components/examples/SignupSuccess"
        '400':
          "$ref": "#/components/responses/BadRequestResponse"
        '409':
          "$ref": "#/components/responses/ConflictResponse"
  "/api/orders":
    get:
      tags:
      - Orders
      summary: 주문 목록 조회
      description: buyerUsername, status 조건으로 주문을 조회합니다. searchMode=DERIVED는 파생 쿼리,
        searchMode=JPQL은 nullable 조건 JPQL을 사용합니다.
      operationId: getOrders
      parameters:
      - name: page
        in: query
        description: 페이지 번호
        required: false
        schema:
          type: integer
          format: int32
          default: 0
        example: 0
      - name: size
        in: query
        description: 페이지 크기
        required: false
        schema:
          type: integer
          format: int32
          default: 10
        example: 10
      - name: buyerUsername
        in: query
        description: 구매자 username 필터
        required: false
        schema:
          type: string
        example: alice
      - name: status
        in: query
        description: 주문 상태 필터
        required: false
        schema:
          type: string
          enum:
          - CREATED
          - PAID
          - SHIPPED
          - CANCELLED
        example: PAID
      - name: searchMode
        in: query
        description: 조회 전략. DERIVED는 파생 쿼리, JPQL은 조건식 기반 JPQL
        required: false
        schema:
          type: string
          enum:
          - DERIVED
          - JPQL
        example: JPQL
      responses:
        '200':
          description: 주문 목록 조회 성공
          content:
            application/json:
              examples:
                success:
                  "$ref": "#/components/examples/OrderListSuccess"
        '400':
          "$ref": "#/components/responses/BadRequestResponse"
    post:
      tags:
      - Orders
      summary: 주문 생성
      description: 주문과 주문 라인을 함께 생성합니다.
      operationId: createOrder
      requestBody:
        description: 주문 생성 요청 바디
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateOrderRequest"
            examples:
              request:
                "$ref": "#/components/examples/OrderCreateRequest"
        required: true
      responses:
        '201':
          description: 주문 생성 성공
          content:
            application/json:
              examples:
                success:
                  "$ref": "#/components/examples/OrderCreateSuccess"
        '400':
          "$ref": "#/components/responses/BadRequestResponse"
  "/api/orders/{id}/ship":
    post:
      tags:
      - Orders
      summary: 주문 배송
      description: PAID 상태 주문을 SHIPPED 상태로 전이합니다.
      operationId: shipOrder
      parameters:
      - name: id
        in: path
        description: 주문 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 2
      responses:
        '200':
          description: 주문 배송 성공
          content:
            application/json:
              examples:
                success:
                  "$ref": "#/components/examples/OrderShipSuccess"
        '404':
          "$ref": "#/components/responses/NotFoundResponse"
        '409':
          "$ref": "#/components/responses/ConflictResponse"
  "/api/orders/{id}/pay":
    post:
      tags:
      - Orders
      summary: 주문 결제
      description: CREATED 상태 주문을 결제합니다. PG 승인 후 주문 완료 저장이 일시 실패하면 처리 중 상태와 조회 URL을 반환합니다.
      operationId: payOrder
      parameters:
      - name: id
        in: path
        description: 주문 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 2
      responses:
        '200':
          description: 주문 결제 성공
          content:
            application/json:
              examples:
                success:
                  "$ref": "#/components/examples/OrderPaySuccess"
        '202':
          description: 결제 승인 후 주문 처리 중
          content:
            application/json:
              examples:
                processing:
                  "$ref": "#/components/examples/OrderPayProcessing"
        '404':
          "$ref": "#/components/responses/NotFoundResponse"
        '409':
          "$ref": "#/components/responses/ConflictResponse"
  "/api/orders/{id}/cancel":
    post:
      tags:
      - Orders
      summary: 주문 취소
      description: CREATED 또는 PAID 상태 주문을 CANCELLED 상태로 전이합니다.
      operationId: cancelOrder
      parameters:
      - name: id
        in: path
        description: 주문 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 2
      responses:
        '200':
          description: 주문 취소 성공
          content:
            application/json:
              examples:
                success:
                  "$ref": "#/components/examples/OrderCancelSuccess"
        '404':
          "$ref": "#/components/responses/NotFoundResponse"
        '409':
          "$ref": "#/components/responses/ConflictResponse"
  "/api/orders/{id}":
    get:
      tags:
      - Orders
      summary: 주문 단건 조회
      description: 상세 조회는 EntityGraph 기반으로 buyer와 orderLines를 함께 로드합니다.
      operationId: getOrder
      parameters:
      - name: id
        in: path
        description: 주문 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        '200':
          description: 주문 상세 조회 성공
          content:
            application/json:
              examples:
                success:
                  "$ref": "#/components/examples/OrderDetailSuccess"
        '404':
          "$ref": "#/components/responses/NotFoundResponse"
  "/api/orders/status-summary":
    get:
      tags:
      - Orders
      summary: 주문 상태 요약 조회
      description: 주문 상태별 개수를 projection 기반으로 조회합니다. buyerUsername, status 조건으로 집계
        범위를 줄일 수 있습니다.
      operationId: getOrderStatusSummaries
      parameters:
      - name: buyerUsername
        in: query
        description: 구매자 username 필터
        required: false
        schema:
          type: string
        example: alice
      - name: status
        in: query
        description: 특정 주문 상태만 집계
        required: false
        schema:
          type: string
          enum:
          - CREATED
          - PAID
          - SHIPPED
          - CANCELLED
        example: PAID
      responses:
        '200':
          description: 주문 상태 요약 조회 성공
          content:
            application/json:
              examples:
                success:
                  "$ref": "#/components/examples/OrderStatusSummarySuccess"
        '400':
          "$ref": "#/components/responses/BadRequestResponse"
components:
  schemas:
    SignupRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 10
          minLength: 4
          pattern: "^[a-z0-9]+$"
        password:
          type: string
          maxLength: 15
          minLength: 8
          pattern: "^[a-zA-Z0-9]+$"
      required:
      - password
      - username
    CreateOrderItemRequest:
      type: object
      properties:
        productName:
          type: string
        quantity:
          type: integer
          format: int32
        unitPrice:
          type: number
      required:
      - productName
      - quantity
      - unitPrice
    CreateOrderRequest:
      type: object
      properties:
        buyerUsername:
          type: string
        shippingAddress:
          "$ref": "#/components/schemas/ShippingAddressRequest"
        items:
          type: array
          items:
            "$ref": "#/components/schemas/CreateOrderItemRequest"
      required:
      - buyerUsername
      - items
      - shippingAddress
    ShippingAddressRequest:
      type: object
      properties:
        recipient:
          type: string
        zipCode:
          type: string
        address1:
          type: string
        address2:
          type: string
      required:
      - address1
      - address2
      - recipient
      - zipCode
  responses:
    BadRequestResponse:
      description: 잘못된 요청
      content:
        application/json:
          examples:
            invalidRequest:
              "$ref": "#/components/examples/InvalidRequestFailure"
            signupValidation:
              "$ref": "#/components/examples/SignupValidationFailure"
            invalidOrderItem:
              "$ref": "#/components/examples/InvalidOrderItemFailure"
    NotFoundResponse:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          examples:
            orderNotFound:
              "$ref": "#/components/examples/OrderNotFoundFailure"
    ConflictResponse:
      description: 상태 충돌 또는 중복 요청
      content:
        application/json:
          examples:
            userAlreadyExists:
              "$ref": "#/components/examples/SignupAlreadyExistsFailure"
            invalidOrderTransition:
              "$ref": "#/components/examples/OrderConflictFailure"
            optimisticLockingConflict:
              "$ref": "#/components/examples/OptimisticLockFailure"
  examples:
    SignupRequest:
      summary: 회원가입 요청
      value:
        username: user1
        password: password1
    SignupSuccess:
      summary: 회원가입 성공
      value:
        result: success
        code: '201'
        message: Created
        timestamp: '2026-05-08T13:30:00'
        data:
          username: user1
    SignupValidationFailure:
      summary: 회원가입 유효성 실패
      value:
        result: failure
        code: '400'
        message: Invalid Request
        timestamp: '2026-05-08T13:30:00'
        errors:
          username: Username must be between 4 and 10 characters
          password: Password must be between 8 and 15 characters
    SignupAlreadyExistsFailure:
      summary: 이미 존재하는 사용자
      value:
        result: failure
        code: '409'
        message: user already exists
        timestamp: '2026-05-08T13:30:00'
        errors: 
    OrderCreateRequest:
      summary: 주문 생성 요청
      value:
        buyerUsername: alice
        shippingAddress:
          recipient: Alice Kim
          zipCode: '06236'
          address1: Seoul Gangnam-daero 123
          address2: 10F
        items:
        - productName: Mechanical Keyboard
          quantity: 1
          unitPrice: 129000.0
        - productName: Palm Rest
          quantity: 1
          unitPrice: 25000.0
    OrderStatusSummarySuccess:
      summary: 주문 상태 요약 조회 성공
      value:
        result: success
        code: '200'
        message: Success
        timestamp: '2026-05-08T13:00:00'
        data:
        - status: PAID
          count: 2
        - status: SHIPPED
          count: 1
    OrderListSuccess:
      summary: 주문 목록 조회 성공
      value:
        result: success
        code: '200'
        message: Success
        timestamp: '2026-05-08T13:00:00'
        data:
          content:
          - id: 2
            version: 1
            buyerUsername: alice
            status: PAID
            totalAmount: 154000.0
            paidAt: '2026-05-08T12:05:00'
            shippedAt:
            cancelledAt:
            createdAt: '2026-05-08T11:00:00'
          pageable:
            pageNumber: 0
            pageSize: 10
          totalElements: 1
          totalPages: 1
    OrderDetailSuccess:
      summary: 주문 상세 조회 성공
      value:
        result: success
        code: '200'
        message: Success
        timestamp: '2026-05-08T13:00:00'
        data:
          id: 2
          version: 1
          buyerUsername: alice
          status: PAID
          recipient: Alice Kim
          zipCode: '06236'
          address1: Seoul Gangnam-daero 123
          address2: 10F
          totalAmount: 154000.0
          items:
          - productName: Mechanical Keyboard
            quantity: 1
            unitPrice: 129000.0
            lineAmount: 129000.0
          - productName: Palm Rest
            quantity: 1
            unitPrice: 25000.0
            lineAmount: 25000.0
          paidAt: '2026-05-08T12:05:00'
          shippedAt: 
          cancelledAt: 
          createdAt: '2026-05-08T11:00:00'
    OrderCreateSuccess:
      summary: 주문 생성 성공
      value:
        result: success
        code: '201'
        message: Created
        timestamp: '2026-05-08T13:00:00'
        data:
          id: 5
          version: 0
          buyerUsername: alice
          status: CREATED
          recipient: Alice Kim
          zipCode: '06236'
          address1: Seoul Gangnam-daero 123
          address2: 10F
          totalAmount: 154000.0
          items:
          - productName: Mechanical Keyboard
            quantity: 1
            unitPrice: 129000.0
            lineAmount: 129000.0
          - productName: Palm Rest
            quantity: 1
            unitPrice: 25000.0
            lineAmount: 25000.0
          paidAt: 
          shippedAt: 
          cancelledAt: 
          createdAt: '2026-05-08T13:00:00'
    OrderPaySuccess:
      summary: 주문 결제 성공
      value:
        result: success
        code: '200'
        message: 결제 및 주문이 완료되었습니다.
        timestamp: '2026-05-08T13:00:00'
        data:
          status: PAID
          message: 결제 및 주문이 완료되었습니다.
          pollingUrl: /api/orders/2
          order:
            id: 2
            version: 2
            buyerUsername: alice
            status: PAID
            recipient: Alice Kim
            zipCode: '06236'
            address1: Seoul Gangnam-daero 123
            address2: 10F
            totalAmount: 154000.0
            items:
            - productName: Mechanical Keyboard
              quantity: 1
              unitPrice: 129000.0
              lineAmount: 129000.0
            - productName: Palm Rest
              quantity: 1
              unitPrice: 25000.0
              lineAmount: 25000.0
            paidAt: '2026-05-08T12:05:00'
            paymentCompletionPendingAt:
            shippedAt:
            cancelledAt:
            paymentKey: MOCK-PG-abc123
            createdAt: '2026-05-08T11:00:00'
    OrderPayProcessing:
      summary: 결제 승인 후 주문 처리 중
      value:
        result: success
        code: '202'
        message: 결제는 승인되었고 주문을 처리 중입니다. 잠시 후 주문 내역에서 확인할 수 있습니다.
        timestamp: '2026-05-08T13:00:01'
        data:
          status: PROCESSING
          message: 결제는 승인되었고 주문을 처리 중입니다. 잠시 후 주문 내역에서 확인할 수 있습니다.
          pollingUrl: /api/orders/2
          order:
            id: 2
            version: 2
            buyerUsername: alice
            status: PAYMENT_COMPLETION_PENDING
            recipient: Alice Kim
            zipCode: '06236'
            address1: Seoul Gangnam-daero 123
            address2: 10F
            totalAmount: 154000.0
            items:
            - productName: Mechanical Keyboard
              quantity: 1
              unitPrice: 129000.0
              lineAmount: 129000.0
            - productName: Palm Rest
              quantity: 1
              unitPrice: 25000.0
              lineAmount: 25000.0
            paidAt:
            paymentCompletionPendingAt: '2026-05-08T13:00:01'
            shippedAt:
            cancelledAt:
            paymentKey: MOCK-PG-abc123
            createdAt: '2026-05-08T11:00:00'
    OrderShipSuccess:
      summary: 주문 배송 성공
      value:
        result: success
        code: '200'
        message: Success
        timestamp: '2026-05-08T13:00:00'
        data:
          id: 2
          version: 3
          buyerUsername: alice
          status: SHIPPED
          recipient: Alice Kim
          zipCode: '06236'
          address1: Seoul Gangnam-daero 123
          address2: 10F
          totalAmount: 154000.0
          items:
          - productName: Mechanical Keyboard
            quantity: 1
            unitPrice: 129000.0
            lineAmount: 129000.0
          - productName: Palm Rest
            quantity: 1
            unitPrice: 25000.0
            lineAmount: 25000.0
          paidAt: '2026-05-08T12:05:00'
          shippedAt: '2026-05-08T12:30:00'
          cancelledAt: 
          createdAt: '2026-05-08T11:00:00'
    OrderCancelSuccess:
      summary: 주문 취소 성공
      value:
        result: success
        code: '200'
        message: Success
        timestamp: '2026-05-08T13:00:00'
        data:
          id: 2
          version: 2
          buyerUsername: alice
          status: CANCELLED
          recipient: Alice Kim
          zipCode: '06236'
          address1: Seoul Gangnam-daero 123
          address2: 10F
          totalAmount: 154000.0
          items:
          - productName: Mechanical Keyboard
            quantity: 1
            unitPrice: 129000.0
            lineAmount: 129000.0
          - productName: Palm Rest
            quantity: 1
            unitPrice: 25000.0
            lineAmount: 25000.0
          paidAt: '2026-05-08T12:05:00'
          shippedAt: 
          cancelledAt: '2026-05-08T12:10:00'
          createdAt: '2026-05-08T11:00:00'
    InvalidRequestFailure:
      summary: 쿼리 파라미터 변환 실패
      value:
        result: failure
        code: '400'
        message: Invalid Request
        timestamp: '2026-05-08T13:00:00'
        errors:
          status: Failed to convert value to OrderStatus
    InvalidOrderItemFailure:
      summary: 잘못된 주문 항목
      value:
        result: failure
        code: '400'
        message: unit price must be greater than zero
        timestamp: '2026-05-08T13:00:00'
        errors: 
    OrderNotFoundFailure:
      summary: 주문을 찾을 수 없음
      value:
        result: failure
        code: '404'
        message: Order not found with id 999
        timestamp: '2026-05-08T13:00:00'
        errors: 
    OrderConflictFailure:
      summary: 허용되지 않는 주문 상태 전이
      value:
        result: failure
        code: '409'
        message: 'only created or paid orders can be cancelled. current status: SHIPPED'
        timestamp: '2026-05-08T13:00:00'
        errors: 
    OptimisticLockFailure:
      summary: 동시 수정 충돌
      value:
        result: failure
        code: '409'
        message: order was modified concurrently. retry the request
        timestamp: '2026-05-08T13:00:00'
        errors: 